alterei o tipo de variável de população para unsigned long int

+++ CartasSuperTrunfo.c
@@ -6,7 +6,7 @@ int main(){
     char estado1, estado2;
     char codigo1[4], codigo2[4];
     char cidade1[20], cidade2[20];
-    int populacao1, populacao2;
+    unsigned long int populacao1, populacao2;
     float area1, area2;
     float pib1, pib2;
     int ptur1, ptur2;
@@ -24,7 +24,7 @@ int main(){
     scanf(" %[^\n]s", cidade1); // [^\n] permite identificar o espaço no nome da cidade
 
     printf("Digite a população da cidade da Carta 1: ");
-    scanf("%d", &populacao1);
+    scanf("%lu", &populacao1);
 
     printf("Digite a área da cidade da Carta 1: ");
     scanf("%f", &area1);
@@ -48,7 +48,7 @@ int main(){
     scanf(" %[^\n]s", cidade2); // [^\n] permite identificar o espaço no nome da cidade
 
     printf("Digite a população da cidade da Carta 2: ");
-    scanf("%d", &populacao2);
+    scanf("%lu", &populacao2);
 
     printf("Digite a área da cidade da Carta 2: ");
     scanf("%f", &area2);
@@ -64,8 +64,8 @@ int main(){
 // CALCULA A DENSIDADE POPULACIONAL E O PIB PER CAPITA DAS CIDADES
     denpop1 = (float) populacao1 / area1;
     denpop2 = (float) populacao2 / area2;
-    pibpc1 = pib1 / (float) populacao1 *1000000000; //multiplicado por bilhões de reais
-    pibpc2 = pib2 / (float) populacao2 *1000000000; //multiplicado por bilhões de reais
+    pibpc1 = (pib1 / (float) populacao1) *1000000000; //multiplicado por bilhões de reais
+    pibpc2 = (pib2 / (float) populacao2) *1000000000; //multiplicado por bilhões de reais
 
 // CALCULA SUPER PODER
     float superpoder1 = (float) populacao1 + area1 + pib1 + (float) ptur1 + pibpc1 + (1 / denpop1);
@@ -76,7 +76,7 @@ int main(){
     printf("Estado: %c\n", estado1);
     printf("Código: %s\n", codigo1);
     printf("Nome da Cidade: %s\n", cidade1);
-    printf("População: %d\n", populacao1);
+    printf("População: %lu\n", populacao1);
     printf("Área: %.2f km²\n", area1);
     printf("PIB: %.2f bilhões de reais\n", pib1);
     printf("Número de Pontos Turísticos: %d\n", ptur1);
@@ -91,7 +91,7 @@ int main(){
     printf("Estado: %c\n", estado2);
     printf("Código: %s\n", codigo2);
     printf("Nome da Cidade: %s\n", cidade2);
-    printf("População: %d\n", populacao2);
+    printf("População: %lu\n", populacao2);
     printf("Área: %.2f km²\n", area2);
     printf("PIB: %.2f bilhões de reais\n", pib2);
     printf("Número de Pontos Turísticos: %d\n", ptur2);
